fix mm reload after import
