+++ src/frontend/screens/Onboarding/walletSelection/index.tsx
@@ -133,7 +133,6 @@ const WalletSelection: React.FC<WalletSelectionProps> = function (props) {
         console.error('There was a problem importing MetaMask!')
         return
       }
-      connectMetaMaskExtension()
     }
   }
